requirements.txt
line-bot-sdk==2.1.0
chainer==7.8.1
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-auth==2.5.0
grpcio>=1.29.0
requests==2.27.1
